requirements.txt
autopep8==1.6.0
awscli==1.25.45
boto3==1.24.33
coloredlogs==15.0.1
django-environ==0.9.0
djlint==0.3.4
Flask-Navigation==0.2.0
Flask==2.1.3
mediapipe==0.8.10.1
pandas==1.4.3
pipdeptree==2.2.1
pylint==2.14.5
verboselogs==1.7